Add files via upload
This is my first attempt at creating a TB78 compatible MailExtension. I've updated lots of the code according to the guide at https://developer.thunderbird.net/add-ons/updating/tb78 and I've implemented the BootstrapLoader (https://github.com/thundernest/addon-developer-support/wiki/Using-the-BootstrapLoader-API-to-convert-a-Legacy-Bootstrap-WebExtension-into-a-MailExtension-for-Thunderbird-78).  Currently, this code is failing in bootstrap.js when the startup function does its first Cu.import call.